refactor : (#42) remove unnecessary imports
