+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/auth/command/dto/response/GithubAccessTokenResponse.kt
@@ -1,4 +1,4 @@
-package entry.dsm.gitauth.equusgithubauth.domain.auth.presentation.response
+package entry.dsm.gitauth.equusgithubauth.domain.auth.command.dto.response
 
 data class GithubAccessTokenResponse(
     val githubAccessToken: String,
