+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/auth/command/service/LogoutService.kt
@@ -1,6 +1,7 @@
-package entry.dsm.gitauth.equusgithubauth.domain.auth.service
+package entry.dsm.gitauth.equusgithubauth.domain.auth.command.service
 
 
+import entry.dsm.gitauth.equusgithubauth.domain.auth.command.service.component.TokenBlackListService
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.repository.RefreshTokenRepository
 import entry.dsm.gitauth.equusgithubauth.global.security.jwt.JwtTokenProvider
 import entry.dsm.gitauth.equusgithubauth.global.security.jwt.exception.JwtTokenInvalidException
