+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/user/service/UserService.kt
@@ -1,8 +1,7 @@
 package entry.dsm.gitauth.equusgithubauth.domain.user.service
 
-import entry.dsm.gitauth.equusgithubauth.domain.auth.service.ValidateGithubOrganizationService
+import entry.dsm.gitauth.equusgithubauth.domain.auth.command.service.ValidateGithubOrganizationService
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.User
-import entry.dsm.gitauth.equusgithubauth.domain.user.entity.enums.UserRole
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.repository.UserRepository
 import entry.dsm.gitauth.equusgithubauth.domain.user.presentation.dto.response.LoginSuccessResponse
 import entry.dsm.gitauth.equusgithubauth.global.external.github.presentation.controller.GithubApiClient
